# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pollo Charly API",
        "description": "Documentación oficial interactiva de los endpoints RESTful del backend de Pollo Charly.",
        "contact": {
            "name": "Equipo de Desarrollo Pollo Charly"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor Actual (Relativo)"
        },
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local (localhost:8000)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local (127.0.0.1:8000)"
        }
    ],
    "paths": {
...
        "/api/dishes": {
            "get": {
                "tags": [
                    "Gestión de Menú"
                ],
                "summary": "Listar platillos del menú",
                "description": "Obtiene el listado paginado de platillos registrados, permitiendo filtrar por búsqueda de texto, categoría, menú del día y estado activo.",
                "operationId": "listDishes",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre o descripción",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrar por ID de categoría",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar por estado activo/inactivo",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "is_daily_menu",
                        "in": "query",
                        "description": "Filtrar platillos del menú del día",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de elementos por página (por defecto 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de platillos obtenido correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DishResource"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gestión de Menú"
                ],
                "summary": "Crear un nuevo platillo con receta y complementos",
                "description": "Registra un nuevo platillo en el menú, definiendo opcionalmente su receta de insumos con cantidades por porción y su lista de complementos aplicables.",
                "operationId": "createDish",
                "requestBody": {
                    "description": "Datos del nuevo platillo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDishRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Platillo registrado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Platillo registrado exitosamente junto con su receta y complementos."
                                        },
                                        "dish": {
                                            "$ref": "#/components/schemas/DishResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (nombre duplicado, precio inválido, etc.).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El nombre del platillo ya existe."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado (requiere rol Administrador).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No tienes permisos para acceder a este módulo. Se requiere rol de administradora."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DishCategoryResource": {
                "title": "Dish Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pollos"
                    }
                },
                "type": "object"
            },
            "DishRecipeItemResource": {
                "title": "Dish Recipe Item",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "supply_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "supply_name": {
                        "type": "string",
                        "example": "Pechuga de Pollo"
                    },
                    "measurement_unit": {
                        "type": "string",
                        "example": "kg"
                    },
                    "required_quantity": {
                        "type": "number",
                        "format": "float",
                        "example": 0.5
                    }
                },
                "type": "object"
            },
...
            "DishResource": {
                "title": "Dish Resource",
                "description": "Información completa del platillo con su receta y complementos asociados",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pollo Frito Familiar"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/DishCategoryResource"
                            }
                        ],
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "8 piezas de pollo frito crujiente",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 125
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://cdn.pollocharly.com/dishes/familiar.jpg",
                        "nullable": true
                    },
                    "is_daily_menu": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "recipes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DishRecipeItemResource"
                        }
                    },
                    "complements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DishComplementItemResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-09-20T12:00:00.000000Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese su token de acceso Bearer generado en /api/login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}